+++ packages/tools/examples/stackSegmentation/index.ts
@@ -1,10 +1,5 @@
-import {
-  cache,
-  Enums,
-  RenderingEngine,
-  metaData,
-  imageLoader,
-} from '@cornerstonejs/core';
+import { Enums, RenderingEngine, imageLoader } from '@cornerstonejs/core';
+import * as cornerstone from '@cornerstonejs/core';
 import * as cornerstoneTools from '@cornerstonejs/tools';
 import {
   createImageIdsAndCacheMetaData,
@@ -13,6 +8,7 @@ import {
   setTitleAndDescription,
   addButtonToToolbar,
 } from '../../../../utils/demo/helpers';
+import { fillStackSegmentationWithMockData } from '../../../../utils/test/testUtils';
 
 // This is for debugging purposes
 console.warn(
@@ -27,7 +23,6 @@ const {
   StackScrollTool,
   Enums: csToolsEnums,
   RectangleScissorsTool,
-  SphereScissorsTool,
   CircleScissorsTool,
   BrushTool,
   PaintFillTool,
@@ -91,8 +86,6 @@ content.append(instructions);
 const brushInstanceNames = {
   CircularBrush: 'CircularBrush',
   CircularEraser: 'CircularEraser',
-  SphereBrush: 'SphereBrush',
-  SphereEraser: 'SphereEraser',
   ThresholdBrush: 'ThresholdBrush',
   DynamicThreshold: 'DynamicThreshold',
 };
@@ -100,8 +93,6 @@ const brushInstanceNames = {
 const brushStrategies = {
   [brushInstanceNames.CircularBrush]: 'FILL_INSIDE_CIRCLE',
   [brushInstanceNames.CircularEraser]: 'ERASE_INSIDE_CIRCLE',
-  [brushInstanceNames.SphereBrush]: 'FILL_INSIDE_SPHERE',
-  [brushInstanceNames.SphereEraser]: 'ERASE_INSIDE_SPHERE',
   [brushInstanceNames.ThresholdBrush]: 'THRESHOLD_INSIDE_CIRCLE',
   [brushInstanceNames.DynamicThreshold]: 'THRESHOLD_INSIDE_CIRCLE',
 };
@@ -109,8 +100,6 @@ const brushStrategies = {
 const brushValues = [
   brushInstanceNames.CircularBrush,
   brushInstanceNames.CircularEraser,
-  brushInstanceNames.SphereBrush,
-  brushInstanceNames.SphereEraser,
   brushInstanceNames.ThresholdBrush,
   brushInstanceNames.DynamicThreshold,
 ];
@@ -119,7 +108,6 @@ const optionsValues = [
   ...brushValues,
   RectangleScissorsTool.toolName,
   CircleScissorsTool.toolName,
-  SphereScissorsTool.toolName,
   PaintFillTool.toolName,
 ];
 
@@ -272,7 +260,6 @@ function setupTools(toolGroupId) {
   cornerstoneTools.addTool(SegmentationDisplayTool);
   cornerstoneTools.addTool(RectangleScissorsTool);
   cornerstoneTools.addTool(CircleScissorsTool);
-  cornerstoneTools.addTool(SphereScissorsTool);
   cornerstoneTools.addTool(PaintFillTool);
   cornerstoneTools.addTool(BrushTool);
 
@@ -290,7 +277,6 @@ function setupTools(toolGroupId) {
   toolGroup.addTool(SegmentationDisplayTool.toolName);
   toolGroup.addTool(RectangleScissorsTool.toolName);
   toolGroup.addTool(CircleScissorsTool.toolName);
-  toolGroup.addTool(SphereScissorsTool.toolName);
   toolGroup.addTool(PaintFillTool.toolName);
   toolGroup.addToolInstance(
     brushInstanceNames.CircularBrush,
@@ -304,20 +290,6 @@ function setupTools(toolGroupId) {
     BrushTool.toolName,
     {
       activeStrategy: brushStrategies.CircularEraser,
-    }
-  );
-  toolGroup.addToolInstance(
-    brushInstanceNames.SphereBrush,
-    BrushTool.toolName,
-    {
-      activeStrategy: brushStrategies.SphereBrush,
-    }
-  );
-  toolGroup.addToolInstance(
-    brushInstanceNames.SphereEraser,
-    BrushTool.toolName,
-    {
-      activeStrategy: brushStrategies.SphereEraser,
     }
   );
   toolGroup.addToolInstance(
@@ -387,38 +359,6 @@ function setupTools(toolGroupId) {
 // ============================= //
 
 /**
- * Adds two concentric circles to each axial slice of the demo segmentation.
- */
-function createMockEllipsoidSegmentation(imageIds, segmentationImageIds) {
-  const { rows, columns } = metaData.get('imagePlaneModule', imageIds[0]);
-  const dimensions = [columns, rows, imageIds.length];
-
-  const center = [dimensions[0] / 2, dimensions[1] / 2, dimensions[2] / 2];
-  const outerRadius = 64;
-  const innerRadius = 32;
-  for (let z = 0; z < dimensions[2]; z++) {
-    let voxelIndex = 0;
-    const image = cache.getImage(segmentationImageIds[z]);
-    const scalarData = image.getPixelData();
-    for (let y = 0; y < dimensions[1]; y++) {
-      for (let x = 0; x < dimensions[0]; x++) {
-        const distanceFromCenter = Math.sqrt(
-          (x - center[0]) * (x - center[0]) +
-            (y - center[1]) * (y - center[1]) +
-            (z - center[2]) * (z - center[2])
-        );
-        if (distanceFromCenter < innerRadius) {
-          scalarData[voxelIndex] = 1;
-        } else if (distanceFromCenter < outerRadius) {
-          scalarData[voxelIndex] = 2;
-        }
-        voxelIndex++;
-      }
-    }
-  }
-}
-
-/**
  * Runs the demo
  */
 async function run() {
@@ -465,10 +405,11 @@ async function run() {
 
   await viewport.setStack(imageIdsArray, 0);
 
-  createMockEllipsoidSegmentation(
-    imageIdsArray.slice(0, 2),
-    segmentationImageIds
-  );
+  fillStackSegmentationWithMockData({
+    imageIds: imageIdsArray.slice(0, 2),
+    segmentationImageIds,
+    cornerstone,
+  });
 
   renderingEngine.renderViewports([viewportId]);
 
@@ -478,11 +419,9 @@ async function run() {
       representation: {
         type: csToolsEnums.SegmentationRepresentations.Labelmap,
         data: {
-          imageIdReferenceMap: new Map(
-            imageIdsArray.map((imageId, index) => [
-              imageId,
-              segmentationImageIds[index],
-            ])
+          imageIdReferenceMap: cstUtils.segmentation.createImageIdReferenceMap(
+            imageIdsArray,
+            segmentationImageIds
           ),
         },
       },