+++ packages/tools/src/index.ts
@@ -67,6 +67,7 @@ import {
   OverlayGridTool,
   SegmentationIntersectionTool,
   FreehandROISculptorTool,
+  SegmentSelectTool,
 } from './tools';
 
 import VideoRedactionTool from './tools/annotation/VideoRedactionTool';
@@ -131,6 +132,7 @@ export {
   RectangleROIStartEndThresholdTool,
   BrushTool,
   OrientationMarkerTool,
+  SegmentSelectTool,
   // Synchronizers
   synchronizers,
   Synchronizer,