+++ packages/tools/src/stateManagement/annotation/annotationState.ts
@@ -80,18 +80,18 @@ function addAnnotation(
   }
 
   const manager = getAnnotationManager();
-  const groupKey = manager.getGroupKey(annotationGroupSelector);
-
-  manager.addAnnotation(annotation, groupKey);
 
   // if the annotation manager selector is an element, trigger the
   // annotation added event for that element.
   if (annotationGroupSelector instanceof HTMLDivElement) {
+    const groupKey = manager.getGroupKey(annotationGroupSelector);
+    manager.addAnnotation(annotation, groupKey);
     triggerAnnotationAddedForElement(annotation, annotationGroupSelector);
   } else {
     // if no element is provided, render all viewports that have the
     // same frame of reference.
     // Todo: we should do something else here for other types of annotation managers.
+    manager.addAnnotation(annotation);
     triggerAnnotationAddedForFOR(annotation);
   }
 
@@ -126,6 +126,9 @@ function getNumberOfAnnotations(
  * @param annotationUID - The unique identifier for the annotation.
  */
 function removeAnnotation(annotationUID: string): void {
+  if (!annotationUID) {
+    return;
+  }
   const manager = getAnnotationManager();
   const annotation = manager.getAnnotation(annotationUID);