+++ packages/tools/src/utilities/segmentation/createBidirectionalToolData.ts
@@ -1,0 +1,68 @@
+import type { Types } from '@cornerstonejs/core';
+import type { Annotation } from '../../types/AnnotationTypes';
+
+export type BidirectionalData = {
+  majorAxis: [Types.Point3, Types.Point3];
+  minorAxis: [Types.Point3, Types.Point3];
+  maxMajor: number;
+  maxMinor: number;
+  segmentIndex: number;
+  label?: string;
+  color?: string | number[];
+  referencedImageId: string;
+  FrameOfReferenceUID: string;
+};
+
+/**
+ * Creates data suitable for the BidirectionalTool from the basic bidirectional
+ * data object.
+ */
+export default function createBidirectionalToolData(
+  bidirectionalData: BidirectionalData,
+  viewport
+): Annotation {
+  const {
+    majorAxis,
+    minorAxis,
+    label = '',
+    FrameOfReferenceUID,
+    referencedImageId,
+  } = bidirectionalData;
+  const [major0, major1] = majorAxis;
+  const [minor0, minor1] = minorAxis;
+
+  const { viewUp, viewPlaneNormal } = viewport.getCamera();
+  const points = [major0, major1, minor0, minor1];
+  const bidirectionalToolData = {
+    highlighted: true,
+    invalidated: true,
+    metadata: {
+      toolName: 'Bidirectional',
+      viewPlaneNormal,
+      viewUp,
+      FrameOfReferenceUID,
+      referencedImageId,
+    },
+    data: {
+      handles: {
+        points,
+        textBox: {
+          hasMoved: false,
+          worldPosition: [0, 0, 0] as Types.Point3,
+          worldBoundingBox: {
+            topLeft: [0, 0, 0] as Types.Point3,
+            topRight: [0, 0, 0] as Types.Point3,
+            bottomLeft: [0, 0, 0] as Types.Point3,
+            bottomRight: [0, 0, 0] as Types.Point3,
+          },
+        },
+        activeHandleIndex: null,
+      },
+      label,
+      cachedStats: {},
+    },
+    isLocked: false,
+    isVisible: true,
+  };
+  return bidirectionalToolData;
+}